Solution to Simple Calculator

+++ Simple Calculator@@6788dd63490637cfc44e1ba8/code.c
@@ -1,6 +1,6 @@
 #include <stdio.h>
 int main() {
-    float a,b;
+    int a,b;
     char ch;
     scanf("%d %d",&a,&b);
     scanf(" %c",&ch);
